[ADD] New system init
