+++ src/__init__.py
@@ -1,2 +1,2 @@
 from . import policy
-from . import _system
+from . import old_system
